@@ -12,25 +12,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
+    "@angular/animations": "6.1.4",
+    "@angular/common": "6.1.4",
+    "@angular/compiler": "6.1.4",
+    "@angular/core": "6.1.4",
+    "@angular/forms": "6.1.4",
+    "@angular/http": "6.1.4",
+    "@angular/platform-browser": "6.1.4",
+    "@angular/platform-browser-dynamic": "6.1.4",
+    "@angular/router": "6.1.4",
     "bootstrap": "^4.0.0-beta",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
-    "rxjs": "^5.4.2",
-    "zone.js": "^0.8.4"
+    "rxjs": "^6.2.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.5",
+    "@angular/compiler-cli": "6.1.4",
+    "@angular/language-service": "6.1.4",
     "@reactivex/rxjs": "^5.5.0",
     "@types/jasmine": "2.8.6",
     "@types/node": "~10.0.2",
@@ -46,6 +47,6 @@
     "protractor": "~5.3.1",
     "ts-node": "~6.0.2",
     "tslint": "~5.9.1",
-    "typescript": "^2.6.2"
+    "typescript": "2.9.2"
   }
 }
