@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from "rxjs/Observable";
+/* import { Observable } from "rxjs/Observable"; */
 import { Http, Response } from '@angular/http';
 
 import 'rxjs/add/operator/map';
@@ -22,7 +22,7 @@ export class InfoAggregateService {
 
   constructor(private http: Http) { }
 
-  public getInfoAggregate(): Observable<InfoAggregate> {
+  /* public getInfoAggregate(): Observable<InfoAggregate> {
     return this.http
       .get(API_URL + '/infoAggregate')
       .map(response => response.json().infoAggregate.map(r => {
@@ -35,5 +35,5 @@ export class InfoAggregateService {
   private handleError(error: Response | any) {
     console.error('ApiService::handleError', error);
     return Observable.throw(error);
-  }
+  } */
 }
